fixed all ml-integration for prediction

+++ backend/sgms-backend/src/main/java/com/sgms/sgms_backend/dto/Department/FloorResponse.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @AllArgsConstructor
 public class FloorResponse {
 
-    private Long floorId;
+    private Integer floorId;
 
     private String floorNumber;
 }
